MOD: create data set for transfer learning

+++ transfer_learning/create_train_small_set.py
@@ -18,8 +18,18 @@ checkDir('train_s/data')
 checkDir('train_s/data/dog/')
 checkDir('train_s/data/cat/')
 
-for i in range(0, 750):
-   name = 'dog.'+str(i)+'.jpg'
-   shutil.copyfile('../train/'+name, 'train_s/data/dog/'+name)
-   name = 'cat.'+str(i)+'.jpg'
-   shutil.copyfile('../train/'+name, 'train_s/data/cat/'+name)
+num_dog = 0
+num_cat = 0
+for i in range(0, 5000):
+    name = 'dog.'+str(i)+'.jpg'
+    if num_dog < 750 and os.path.exists('../train/'+name):
+        shutil.copyfile('../train/'+name, 'train_s/data/dog/'+name)
+        num_dog += 1
+      
+    name = 'cat.'+str(i)+'.jpg'
+    if num_cat < 750 and os.path.exists('../train/'+name):
+        shutil.copyfile('../train/'+name, 'train_s/data/cat/'+name)
+        num_cat += 1
+        
+    if num_dog >= 750 and num_cat >= 750:
+        break
